requirements.txt
Django==2.0.6
django-bootstrap3==11.0.0
django-paypal==0.5.0
django-smtp-ssl==1.0
django-widget-tweaks==1.4.3
easy-thumbnails==2.5
mysqlclient==1.3.12
pagseguro==0.3.4
Pillow==5.3.0
djangorestframework==3.10.3
